refactor: 멀티 모듈 구조에 맞춰 domain 레이어 분리
