+++ nmnb-domain/src/main/kotlin/nmnb/domain/like/Like.kt
@@ -1,7 +1,5 @@
-package cv.nmnb.domain.post.domain
+package nmnb.domain.like
 
-import cv.nmnb.domain.user.domain.User
-import cv.nmnb.global.common.domain.BaseEntity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
@@ -11,24 +9,27 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
+import jakarta.persistence.UniqueConstraint
+import nmnb.domain.BaseEntity
+import nmnb.domain.post.Post
+import nmnb.domain.user.User
 
 @Entity
-@Table(name = "posts")
-class Post(
-    @Column(nullable = false)
-    val url: String,
-
-    @Column(nullable = false)
-    val thumbnailUrl: String,
-
-    var description: String? = null,
-
+@Table(
+    name = "user_post_likes",
+    uniqueConstraints = [UniqueConstraint(columnNames = ["user_id", "post_id"])],
+)
+class Like(
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    val user: User,
+    val likedBy: User,
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    val post: Post,
 ) : BaseEntity() {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_id")
+    @Column(name = "like_id")
     val id: Long? = null
 }
