+++ nmnb-domain/src/main/kotlin/nmnb/domain/post/Post.kt
@@ -1,0 +1,34 @@
+package nmnb.domain.post
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.FetchType
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.ManyToOne
+import jakarta.persistence.Table
+import nmnb.domain.BaseEntity
+import nmnb.domain.user.User
+
+@Entity
+@Table(name = "posts")
+class Post(
+    @Column(nullable = false)
+    val url: String,
+
+    @Column(nullable = false)
+    val thumbnailUrl: String,
+
+    var description: String? = null,
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    val user: User,
+) : BaseEntity() {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "post_id")
+    val id: Long? = null
+}
