+++ nmnb-domain/src/main/kotlin/nmnb/domain/user/PetOwnershipStatus.kt
@@ -1,4 +1,4 @@
-package cv.nmnb.domain.user.domain
+package nmnb.domain.user
 
 enum class PetOwnershipStatus {
     HAS_PET,
