+++ nmnb-domain/src/main/kotlin/nmnb/domain/user/generator/annotation/UserId.kt
@@ -1,6 +1,6 @@
 package cv.nmnb.global.generator.annotation
 
-import cv.nmnb.global.generator.CustomIdGenerator
+import nmnb.domain.user.generator.CustomIdGenerator
 import org.hibernate.annotations.IdGeneratorType
 
 @Target(AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
