+++ nmnb-domain/src/main/resources/application-local.yml
@@ -16,7 +16,6 @@ spring:
     properties:
       hibernate:
         format_sql: true
-
   data:
     redis:
       host: ${LOCAL_REDIS_URL}
