--- src/main/kotlin/cv/nmnb/domain/like/domain/Like.kt
@@ -1,35 +1,0 @@
-package cv.nmnb.domain.like.domain
-
-import cv.nmnb.domain.post.domain.Post
-import cv.nmnb.domain.user.domain.User
-import cv.nmnb.global.common.domain.BaseEntity
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.FetchType
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-import jakarta.persistence.JoinColumn
-import jakarta.persistence.ManyToOne
-import jakarta.persistence.Table
-import jakarta.persistence.UniqueConstraint
-
-@Entity
-@Table(
-    name = "user_post_likes",
-    uniqueConstraints = [UniqueConstraint(columnNames = ["user_id", "post_id"])],
-)
-class Like(
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    val likedBy: User,
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    val post: Post,
-) : BaseEntity() {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "like_id")
-    val id: Long? = null
-}
